pyproject.toml
[project]
name = "azure_jobs"
version = "0.1.13"
description = "Add your description here"
readme = "README.md"
authors = [{ name = "HSPK", email = "whxway@whu.edu.cn" }]
requires-python = ">=3.10"
dependencies = [
    "click>=8.2.1",
    "pyyaml>=6.0.2",
    "rich>=14.3.3",
    "textual>=1.0.0",
    "azure-identity>=1.17.0",
    "requests>=2.31.0",
]

[project.scripts]
aj = "azure_jobs.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.pytest.ini_options]
testpaths = ["tests"]

[dependency-groups]
dev = ["ipykernel>=6.30.1", "pytest>=8.0", "pytest-asyncio>=0.24"]
